requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomysql==0.2.0
aiosignal==1.3.2
aniso8601==10.0.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-WTF==1.2.2
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==9.1.0
nltk==3.9.1
propcache==0.2.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.37
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
zipp==3.21.0
